Delete client-side filtering implementation
The more I think about, the less it makes sense to handle the filtering of events client-side. Imagine a dataset of a 1000+ upcoming events. These would've to load into the client before any filtering gets done or the results of said filtering wouldn't make any sense. This comes with an enormous cost of bandwidth. If you decide to only load a subset, you'd also only be able to filter this subset. The sane approach is to make requests with the filtering options as query params and update the UI with response from a plugin-specific api-endpoint. This way we can always filter through all events without the expense of having them loaded client-side while permitting to load just a subset of, say.. 20 events because more can't be displayed anyway + we can implement a relevancy mechanism for sorting the api response.